MINIFICPP-1900 Make python scripting extension compile on Windows
- Fix compile errors:
    * do the sol2 patch in binary mode
    * put gcc-specific pragmas and attributes in #def's
    * undef COMPILER as some include #defines COMPILER as the compiler version
    * fix the usual Windows filesystem::path -> string stuff
- Fix some warnings
- Enable the script extension in the Windows CI build
- Set up Python in the Windows CI job
- Fix the unit tests

Closes #1392

+++ libminifi/include/io/StreamPipe.h
@@ -37,7 +37,7 @@ inline int64_t pipe(io::InputStream& src, io::OutputStream& dst) {
     if (io::isError(readRet)) return -1;
     if (readRet == 0) break;
     auto remaining = readRet;
-    int transferred = 0;
+    size_t transferred = 0;
     while (remaining > 0) {
       const auto writeRet = dst.write(gsl::make_span(buffer).subspan(transferred, remaining));
       // TODO(adebreceni):
